+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- expose `Path_condition.pp`
+
 ## 0.1 - 2026-02-09
 
 - first release